% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ezrmANOVAoutDF.R
\name{add_dz_to_ANOVA_table}
\alias{add_dz_to_ANOVA_table}
\title{add_dz_to_ANOVA_table}
\usage{
add_dz_to_ANOVA_table(no_dz_anova, nwithin)
}
\description{
add_dz_to_ANOVA_table
}
\author{
Thirsa Huisman
}
