% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvAGO_glmnet.R
\name{cvAGO.glmnet}
\alias{cvAGO.glmnet}
\title{cvAGO.glmnet}
\usage{
cvAGO.glmnet(
  dataset,
  X,
  Y,
  family,
  standardize = TRUE,
  alpha = 1,
  nfolds,
  nCVruns,
  idVarString,
  type.measure,
  ...
)
}
\arguments{
\item{dataset}{this is the dataset}

\item{X}{X design matrix for glmnet}

\item{Y}{Y variable for glmnet}

\item{family}{distribution family for glm}

\item{standardize}{specifies if vars are standardized (default = TRUE)}

\item{alpha}{value of alpha parameter in the elastic net (set to 1 by default = Lasso)}

\item{nfolds}{number of folds per CV run}

\item{nCVruns}{number of CV runs}

\item{idVarString}{variable in dataset identifying the unit of observation in groupdata2's foldid}

\item{type.measure}{measure used for evaluating the GOF}

\item{...}{}
}
\value{
cvfitList= List of cross-validated glmnet models (one per CV run). coefLassodf = dataframe containing the glmnet parameter estimates etc. per CV run
}
\description{
cvAGO.glmnet
}
\examples{
To do
}
