% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvGroupRep_glmnet.R
\name{cvGroupRep.glmnet}
\alias{cvGroupRep.glmnet}
\title{cvGroupRep.glmnet. Runs a by-group repeated cross-validation with glmnet}
\usage{
cvGroupRep.glmnet(
  dataset,
  glmFormulaString,
  family,
  standardize = TRUE,
  alpha = 1,
  nfolds = 10,
  nCVruns = 1,
  idVarString = NULL,
  type.measure,
  seedX = 4444,
  ...
)
}
\arguments{
\item{dataset}{this is the dataset (data frame)}

\item{glmFormulaString}{regression formula compatible with glmnet}

\item{family}{distribution family for glm}

\item{standardize}{specifies if vars are standardized (default = TRUE)}

\item{alpha}{value of alpha parameter in the elastic net (set to 1 by default = Lasso)}

\item{nfolds}{number of folds per CV run (default = 10)}

\item{nCVruns}{number of CV runs (default = 1)}

\item{idVarString}{variable in dataset identifying the unit of observation in groupdata2's foldid. Must be numeric (default = NULL; no grouping variable)}

\item{type.measure}{measure (loss function) used for evaluating the GOF}

\item{seedX}{seed for random number generator, default = 4444}

\item{...}{optional  parameters passed to cv.glmnet()}
}
\value{
cvfitList= List of cross-validated glmnet models (one per CV run). coefLassodf = dataframe containing the glmnet parameter estimates etc. per CV run
}
\description{
cvGroupRep.glmnet. Runs a by-group repeated cross-validation with glmnet
}
