% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ezrmANOVAoutDF.R
\name{format_ANOVA_table}
\alias{format_ANOVA_table}
\title{format_ANOVA_table}
\usage{
format_ANOVA_table(
  ezANOVAobject,
  add_dz = TRUE,
  participant_column_name = "Participantnr"
)
}
\arguments{
\item{add_dz}{add Cohen's dz to the table (!! valid only for within-subject effects, not checked by function!)}

\item{ezanovaObject}{object created by ezANOVA}
}
\value{
formattable object with formatted ANOVA table
}
\description{
format_ANOVA_table
}
\author{
Thirsa Huisman
}
