% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ezrmANOVAoutDF.R
\name{italic_main_and_subscript}
\alias{italic_main_and_subscript}
\title{italic_main_and_subscript}
\usage{
italic_main_and_subscript(main, subscript = "")
}
\description{
italic_main_and_subscript
}
\author{
Thirsa Huisman
}
