% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Common_conversion_functions.R
\name{MsToKmh}
\alias{MsToKmh}
\title{Convert meters per second to kilometers per hour}
\usage{
MsToKmh(x)
}
\arguments{
\item{x}{Speed in m/s}
}
\value{
Speed in km/h
}
\description{
Convert meters per second to kilometers per hour
}
