% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nestByAddGroupKey.R
\name{nestByAddGroupKey}
\alias{nestByAddGroupKey}
\title{nestByAddGroupKey}
\usage{
nestByAddGroupKey(dataframe, groupVars)
}
\arguments{
\item{dataframe}{dataframe to be nested}

\item{groupVars}{grouping variable as character string}
}
\value{
data.frame containing nested data
}
\description{
nests a dataframe by grouping vars “groupVars”
(vector of strings, e.g., groupVars=c("Subj","Condition"))
and adds a string variable "groupKey" indicating the group vars combination
to the "data" list object
}
\examples{
iris |> nestByAddGroupKey("Species")


}
