% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Common_plotting_functions.R
\name{plot_basics_jitter}
\alias{plot_basics_jitter}
\title{plot_basics}
\usage{
plot_basics_jitter(
  error_bar_width = 1,
  point_size = 2,
  stroke_size = 1.5,
  line_width = 1,
  errorbar_line_width = 0.75,
  jitter_width = 0,
  jitter_height = 0
)
}
\arguments{
\item{error_bar_width}{width of the error bars}
}
\description{
This is a list of the basic plot elements that are used in most plots. Shifts are done via position_jitter instead of dodge to avoid errorbar width issues
}
\details{
error_bar_width needs to be scaled according to the size of the x_axis, e.g., if x goes from 0 to 10 vs 0 to 1000, you need to scale the width accordingly.
}
\author{
Thirsa Huisman
}
