% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Common_plotting_functions.R
\name{plot_basics_positiondodge}
\alias{plot_basics_positiondodge}
\title{plot_basics}
\usage{
plot_basics_positiondodge(
  error_bar_width = 1,
  point_size = 2,
  stroke_size = 1.5,
  line_width = 1,
  errorbar_line_width = 0.75,
  positiondodge_width = 0
)
}
\arguments{
\item{error_bar_width}{width of the error bars}
}
\description{
This is a list of the basic plot elements that are used in most plots, note that positiondodge can give errorbar width issues when the nr of groups per x changes.
}
\details{
error_bar_width needs to be scaled according to the size of the x_axis, e.g., if x goes from 0 to 10 vs 0 to 1000, you need to scale the width accordingly.
}
\author{
Thirsa Huisman
}
