% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotThemeAGO.R
\name{plotThemeAGO}
\alias{plotThemeAGO}
\title{Complete plot theme for publication-ready data plots}
\usage{
plotThemeAGO(base_size = 12, base_family = "Arial", ...)
}
\arguments{
\item{base_size}{base size}

\item{base_family}{base font family, default = Arial Unicode MS}

\item{...}{optional  parameters passed to theme()}
}
\description{
Complete plot theme for publication-ready data plots
}
