% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmANOVAcellCheck.R
\name{rmANOVAcellCheck}
\alias{rmANOVAcellCheck}
\title{Check a dataframe prior to running a repeated-measures ANOVA. Detects and removes subjects with empty cells, and issues warning if there are cells containing more than one observation}
\usage{
rmANOVAcellCheck(df, subjIDstr, factorsStr, verbose = F)
}
\arguments{
\item{df}{input dataframe}

\item{subjIDstr}{Name of the column identifying the subjects (string), e.g., "subjCode".}

\item{factorsStr}{Vector of column names of the factors in the rmANOVA (character vector), e.g., c("Condition","Time")}

\item{verbose}{logical (default = FALSE).}
}
\value{
\code{nEmptyCells} = number of empty cells

\code{df_nonEmpty} = data frame where subjects with empty cells are removed

\code{maxObsPerCell} = maximum number of observations in df_nonEmpty per cell (must be 1 for an rmANOVA)
}
\description{
Check a dataframe prior to running a repeated-measures ANOVA. Detects and removes subjects with empty cells, and issues warning if there are cells containing more than one observation
}
