% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Common_plotting_functions.R
\name{show_plot}
\alias{show_plot}
\title{show_plot}
\usage{
show_plot(plot_and_plotname, seperate_plotname = "")
}
\arguments{
\item{plot_and_plotname}{if no seperate_plotname is given, show_plot will try to show "plot_and_plotname" or load "plot_and_plotname.png". Use sepearte_plotname if you save the files to a different folder}

\item{seperate_plotname}{Full name of the plot, eg: "saved_image.png". Needs to be used when figure is stored in subfolder (relative pathing) or other folder entirely (full pathing).}
}
\description{
When knitting it loads the preper saved image, when merely running the code it shows the simple plot.
}
\author{
Thirsa Huisman
}
