% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/streetcrossing.R
\docType{data}
\name{streetcrossing}
\alias{streetcrossing}
\title{Data from Streetcrossing Experiment 2022}
\format{
\subsection{\code{streecrossing}}{

A data frame with 2,771 rows and 10 columns:
\describe{
\item{vp_code}{participant identification code}
\item{label}{Presented car}
\item{modality}{Modality (A = auditory, AV = audiovisual)}
\item{v0}{}
\item{a}{}
\item{gain}{}
\item{track_TTC}{}
\item{nTrials}{Number of trials}
\item{nCross}{velocity at occlusion}

}
}
}
\usage{
streetcrossing
}
\description{
Data from Streetcrossing Experiment 2022
}
\keyword{datasets}
