% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyQuickPsy.R
\name{tidyQuickPsy}
\alias{tidyQuickPsy}
\title{converts a quickpsy object into a tidy tibble}
\usage{
tidyQuickPsy(fit)
}
\arguments{
\item{fit}{quickpsy object}
}
\value{
a list containing the original quickpsy object (fit) and the tibble version (tidy_fit)
}
\description{
converts a quickpsy object into a tidy tibble
}
