% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tukey_outlier_fun2.R
\name{tukey_outlier_fun2}
\alias{tukey_outlier_fun2}
\title{Detect outlier according to the tukey criterion.}
\usage{
tukey_outlier_fun2(data, dv, tukey_crit = 3)
}
\arguments{
\item{dv}{continuous dependent variable AS VECTOR. For TTC experiments insert the estimated TTC here.}

\item{tukey_crit}{factor of the IQR that defines an outlier.}
}
\value{
A list of vectors.
}
\description{
Detect outlier according to the tukey criterion.
}
\examples{
iris |>
dplyr::mutate(tidyr::as_tibble(tukey_outlier_fun(dv=Petal.Length,tukey_crit=3)))


}
