% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ezrmANOVAoutDF.R
\name{ezrmANOVAoutDF}
\alias{ezrmANOVAoutDF}
\title{ezrmANOVAoutDF Function for extracting relevant output of an rmANOVA conducted with ezANOVA to a dataframe}
\usage{
ezrmANOVAoutDF(ezANOVAobj)
}
\arguments{
\item{ezANOVAobj}{the input ezANOVA object}
}
\value{
rmANOVAout: dataframe with relevant statistics per effect
}
\description{
ezrmANOVAoutDF Function for extracting relevant output of an rmANOVA conducted with ezANOVA to a dataframe
}
