% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ezrmANOVAoutDF.R
\name{format_ANOVA_table}
\alias{format_ANOVA_table}
\title{format_ANOVA_table}
\usage{
format_ANOVA_table(
  ezanova_output,
  add_dz = TRUE,
  participant_column_name = "Participantnr"
)
}
\description{
format_ANOVA_table
}
\author{
Thirsa Huisman
}
