% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotQuickPsy.R
\name{plotQuickPsy}
\alias{plotQuickPsy}
\title{plots individual psychometric functions derived from tidy quickpsy data frame.}
\usage{
plotQuickPsy(
  qp,
  values = c("nTrials", "sigmaEst", "se_sigmaEst", "muEst", "se_muEst", "LLRpValue")
)
}
\arguments{
\item{qp}{tidy quickpsy object (see function \code{tidyQuickPsy})}

\item{values}{plotted values (default = \code{c("nTrials", "sigmaEst", "se_sigmaEst", "muEst", "se_muEst", "LLRpValue")})}
}
\value{
returns a list containing two elements:

\code{qp} quickpsy object passed to the function

\code{plot_list} list containing the fitted psychometric function for each participant-condition combination
}
\description{
plots individual psychometric functions derived from tidy quickpsy data frame.
}
