% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotQuickPsy.R
\name{plotQuickPsy}
\alias{plotQuickPsy}
\title{plots individual psychometric functions derived from tidy quickpsy data frame.}
\usage{
plotQuickPsy(
  qp,
  values = c("nTrials", "sigmaEst", "se_sigmaEst", "muEst", "se_muEst", "LLRpValue")
)
}
\arguments{
\item{qp}{tidy quickpsy object (see function tidyQuickPsy)}

\item{values}{plotted values}
}
\value{
list of individual psychometric functions
}
\description{
plots individual psychometric functions derived from tidy quickpsy data frame.
}
