% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotThemeAGO.R
\name{plotThemeAGO}
\alias{plotThemeAGO}
\title{Complete plot theme for TTC estimation plots}
\usage{
plotThemeAGO(base_size = 14, base_family = "Arial Unicode MS")
}
\arguments{
\item{base_size}{base size}

\item{base_family}{base font family, default = Arial Unicode MS}
}
\description{
Complete plot theme for TTC estimation plots
}
