% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Common_plotting_functions.R
\name{pShawe}
\alias{pShawe}
\title{pShawe}
\usage{
pShawe(
  gObj,
  maxwidth = 16,
  maxheight = 25,
  units = "cm",
  dpi = 600,
  device = "png",
  fnSuffix = "",
  imgPath = getwd(),
  ...
)
}
\arguments{
\item{gObj}{ggplot object}
}
\value{
img: the saved image , fName: the filename of the saved image
}
\description{
Saves ggplot object as an image file with DeLuciatoR::ggsave_fitmax, then loads it to display in the RStudio viewer pane.
}
\author{
Daniel Oberfeld-Twistel
}
