% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Common_plotting_functions.R
\name{pShawe}
\alias{pShawe}
\title{pShawe}
\usage{
pShawe(
  gObj,
  maxwidth = 16,
  maxheight = 25,
  units = "cm",
  dpi = 600,
  device = "png",
  fnSuffix = "",
  imgPath = getwd(),
  ...
)
}
\arguments{
\item{gObj}{ggplot object}
}
\value{
the saved then loaded image
}
\description{
Saved image then loads it to ensure that the R image matches the final produced image
}
\author{
Daniel Oberfeld-Twistel
}
