% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tukey.R
\name{tukey}
\alias{tukey}
\title{Detect outlier according to the tukey criterion.}
\usage{
tukey(data, dv, tukey_crit = 3, exclude = F)
}
\arguments{
\item{data}{data frame to apply the tukey outlier analysis on.}

\item{dv}{variable to apply the tukey outlier analysis on.}

\item{tukey_crit}{factor of the IQR that defines an outlier.}

\item{exclude}{logical indicating if outliers should be excluded (default = FALSE).}
}
\value{
data frame including new variables concerning tukey outlier analysis.

\code{trialsInSet} = number of trials in the subset of trials for which the outlier detection is run

\code{IQR} = inter-quartile range

\code{Quant25} = 25\% quantile

\code{Quant75} = 75\% quantile

\code{outlierTukeyLow} = indicates whether the value of dv on a given trial is lower than Tukey_lower_limit (25\% quantile minus tukey_crit* IQR)  (1) or not (0)

\code{outlierTukeyHigh} = indicates whether the value of dv on a given trial is higher than Tukey_upper_limit (75\% quantile plus tukey_crit* IQR)  (1) or not (0)

\code{outlierTukey} = indicates if dv for a given trial exceeds the lower or the higher criterion (1) or is within both criteria (0)

\code{Tukey_lower_limit} = highest value above which values are excluded

\code{Tukey_upper_limit} = lowest value below which values are excluded
}
\description{
Detect outlier according to the tukey criterion.
}
\examples{
tukey(data=iris,dv=Petal.Length,tukey_crit=1.5)

}
