% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tukey.R
\name{tukey}
\alias{tukey}
\title{Detect outlier according to the tukey criterion.}
\usage{
tukey(data, dv, tukey_crit = 3, exclude = F)
}
\arguments{
\item{data}{data frame to apply the tukey outlier analysis on.}

\item{dv}{variable to apply the tukey outlier analysis on.}

\item{tukey_crit}{factor of the IQR that defines an outlier.}

\item{exclude}{logical indicating if outliers should be excluded (default = FALSE).}
}
\value{
data frame including new variables concerning tukey outlier analysis.
}
\description{
Detect outlier according to the tukey criterion.
}
\examples{
tukey(iris,Petal.Length)

}
