% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tukey_outlier_fun.R
\name{tukey_outlier_fun}
\alias{tukey_outlier_fun}
\title{Detect outliers according to a Tukey criterion.}
\usage{
tukey_outlier_fun(dv, tukey_crit = 3)
}
\arguments{
\item{dv}{continuous dependent variable. For TTC experiments insert the estimated TTC here.}

\item{tukey_crit}{factor of the IQR that defines an outlier.}
}
\value{
A list of vectors.
}
\description{
Detect outliers according to a Tukey criterion.
}
\examples{
iris |>
dplyr::mutate(tidyr::as_tibble(tukey_outlier_fun(dv=Petal.Length,tukey_crit=1.5)))


}
