% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tukey_outlier_fun.R
\name{tukey_outlier_fun}
\alias{tukey_outlier_fun}
\title{Detect outliers according to a Tukey criterion.}
\usage{
tukey_outlier_fun(dv, tukey_crit = 3)
}
\arguments{
\item{dv}{continuous dependent variable. For TTC experiments insert the estimated TTC here.}

\item{tukey_crit}{factor of the IQR that defines an outlier.}
}
\value{
A list of vectors:
\itemize{
\item \code{trialsInSet} = total number of trials in the data set
\item \code{IQR} = inter quantile range
\item \code{Quant25} = 25\% quantile
\item \code{Quant75} = 75\% quantile
\item \code{outlierTukeyLow} = indicates if dv for a given trial is lower than the tukey criterion (1) or not (0)
}
}
\description{
Detect outliers according to a Tukey criterion.
}
\examples{
iris |>
dplyr::mutate(tidyr::as_tibble(tukey_outlier_fun(dv=Petal.Length,tukey_crit=1.5)))


}
